Localised multiple kernel k-means

Perform the training step of the localised multiple kernel k-means

lmkkmeans_missingData(Km, parameters, missing = NULL, verbose = FALSE)

Arguments

Km
Array of size N X N X M containing M kernel matrices.

parameters
List of parameters, containing `iteration_count`, `cluster_count`.

missing
Matrix of size N X M containing missingness indicators.

verbose
Boolean flag. If TRUE, at each iteration the iteration number is printed. Defaults to FALSE.

References

Gonen, M. and Margolin, A.A., 2014. Localized data fusion for kernel k-means clustering with application to cancer biology. In Advances in Neural Information Processing Systems (pp. 1305-1313).