Kernel learning integrative clustering

Perform kernel learning integrative clustering

klic(data, M, individualK = NULL, individualMaxK = 6,
  individualClAlgorithm = "kkmeans", globalK = NULL, globalMaxK = 6,
  B = 1000, C = 100, savePlots = FALSE, fileName = "klic")

Arguments

data
List of M datasets, each of size N X P_m, m = 1, ..., M.

M
number of datasets.

individualK
Vector containing the number of clusters in each dataset. Default is NULL. If the number of clusters is not provided, then all the possible values between 2 and individualMaxK are considered and the best value is chosen for each dataset by maximising the silhouette.

individualMaxK
Maximum number of clusters considered for the individual data. Default is 6.

individualClAlgorithm
Clustering algorithm used for clustering of each dataset individually if is required to find the best number of clusters.

globalK
Number of global clusters. Default is NULL. If the number of clusters is not provided, then all the possible values between 2 and globalMaxK are considered and the best value is chosen by maximising the silhouette.

globalMaxK
Maximum number of clusters considered for the final clustering. Default is 6.

B
Number of iterations for consensus clustering. Default is 1000.

C
Maximum number of iterations for localised kernel k-means. Default is 100.

savePlots
If TRUE, a plot of the silhouette is saved in the working folder. Default is FALSE.

fileName
If savePlots is TRUE, this is the name of the png file.

Examples

# Load synthetic data
data1 <- as.matrix(read.csv(system.file("extdata",
"dataset1.csv", package = "klic"), row.names = 1))
data2 <- as.matrix(read.csv(system.file("extdata",
"dataset2.csv", package = "klic"), row.names = 1))
data3 <- as.matrix(read.csv(system.file("extdata",
"dataset3.csv", package = "klic"), row.names = 1))
data <- list(data1, data2, data3)

# Perform clustering with KLIC assuming to know the
# number of clusters in each individual dataset and in
# the final clustering
klicOutput <- klic(data, 3, individualK = c(6,6,6),
globalK = 6, B = 50, C = 5)

# Extract cluster labels
klic_labels <- klicOutput$globalClusterLabels

cluster_labels <- as.matrix(read.csv(system.file("extdata",
"cluster_labels.csv", package = "klic"), row.names = 1))
# Compute ARI
ari <- mclust::adjustedRandIndex(klic_labels, cluster_labels)