Localised multiple kernel k-means

Perform the training step of the localised multiple kernel k-means

lmkkmeans(Km, parameters, verbose = FALSE)

Arguments

Km
kernel matrices

parameters
list of parameters

verbose
Boolean flag. If TRUE, at each iteration the iteration number is printed. Default is FALSE.

References

Gönen, M. and Margolin, A.A., 2014. Localized data fusion for kernel k-means clustering with application to cancer biology. In Advances in Neural Information Processing Systems (pp. 1305-1313).

Examples

# Initialise 100 x 100 x 3 array containing M kernel matrices
# representing three different types of similarities
# between 100 data points
km <- array(NA, c(100, 100, 3))
# Load kernel matrices
km[,,1] <- as.matrix(read.csv(system.file('extdata',
'kernel_matrix1.csv', package = 'klic'), row.names = 1))
km[,,2] <- as.matrix(read.csv(system.file('extdata',
'kernel_matrix2.csv', package = 'klic'), row.names = 1))
km[,,3] <- as.matrix(read.csv(system.file('extdata',
'kernel_matrix3.csv', package = 'klic'), row.names = 1))
#initalize the parameters of the algorithm
parameters <- list()
#set the number of clusters
parameters$cluster_count <- 2
# set the number of iterations
parameters$iteration_count <- 10
#perform training
state <- lmkkmeans(km, parameters)
#display the clustering
print(state$clustering)

#>   [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [38] 1 1 1 1 1 1 1 1 1 1 1 1 1 2 2 2 2 2 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [75] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2

#display the kernel weights
print(state$Theta)

#>             [,1]      [,2]      [,3]
#>   [1,] 0.3346087 0.3307825 0.3346087
#>   [2,] 0.3346710 0.3306580 0.3346710
#>   [3,] 0.3333024 0.3333951 0.3333024
#>   [4,] 0.3343984 0.3312032 0.3343984
#>   [5,] 0.3342298 0.3315404 0.3342298
#>   [6,] 0.3341172 0.3317657 0.3341172
#>   [7,] 0.3338299 0.3323403 0.3338299
#>   [8,] 0.3340767 0.3318466 0.3340767
#>   [9,] 0.3332093 0.3335813 0.3332093
#>  [10,] 0.3341911 0.3316178 0.3341911
#>  [11,] 0.3343757 0.3312487 0.3343757
#>  [12,] 0.3341314 0.3317372 0.3341314
#>  [13,] 0.3341724 0.3316553 0.3341724
#>  [14,] 0.3344422 0.3311155 0.3344422
#>  [15,] 0.3263675 0.3472651 0.3263675
#>  [16,] 0.3342485 0.3315030 0.3342485
#>  [17,] 0.3332488 0.3335023 0.3332488
#>  [18,] 0.3344242 0.3311516 0.3344242
#>  [19,] 0.3344929 0.3310142 0.3344929
#>  [20,] 0.3345346 0.3309309 0.3345346
#>  [21,] 0.2862857 0.4274286 0.2862857
#>  [22,] 0.3341861 0.3316278 0.3341861
#>  [23,] 0.2880555 0.4238890 0.2880555
#>  [24,] 0.3336395 0.3327211 0.3336395
#>  [25,] 0.3342705 0.3314590 0.3342705
#>  [26,] 0.3050784 0.3799726 0.3149490
#>  [27,] 0.3035242 0.3831145 0.3133613
#>  [28,] 0.3036296 0.3832827 0.3130877
#>  [29,] 0.3025953 0.3849131 0.3124917
#>  [30,] 0.3033815 0.3832968 0.3133217
#>  [31,] 0.3047563 0.3808083 0.3144354
#>  [32,] 0.3034524 0.3836614 0.3128862
#>  [33,] 0.3029436 0.3840974 0.3129589
#>  [34,] 0.3036887 0.3831395 0.3131718
#>  [35,] 0.3034048 0.3835057 0.3130895
#>  [36,] 0.3027097 0.3845836 0.3127067
#>  [37,] 0.3040297 0.3823205 0.3136497
#>  [38,] 0.3039653 0.3827957 0.3132390
#>  [39,] 0.3035782 0.3830235 0.3133983
#>  [40,] 0.3027081 0.3845570 0.3127349
#>  [41,] 0.3028365 0.3843065 0.3128570
#>  [42,] 0.3053511 0.3800074 0.3146415
#>  [43,] 0.3034776 0.3834709 0.3130515
#>  [44,] 0.3135599 0.3635897 0.3228504
#>  [45,] 0.3035905 0.3830657 0.3133438
#>  [46,] 0.3096053 0.3715675 0.3188272
#>  [47,] 0.2830834 0.3931296 0.3237870
#>  [48,] 0.3041353 0.3820879 0.3137767
#>  [49,] 0.3028206 0.3844560 0.3127234
#>  [50,] 0.3043331 0.3818838 0.3137830
#>  [51,] 0.3358407 0.3320560 0.3321033
#>  [52,] 0.3359117 0.3319975 0.3320907
#>  [53,] 0.3357570 0.3321992 0.3320438
#>  [54,] 0.3361421 0.3316934 0.3321644
#>  [55,] 0.3357170 0.3320883 0.3321946
#>  [56,] 0.2900307 0.4236407 0.2863285
#>  [57,] 0.3358590 0.3321058 0.3320352
#>  [58,] 0.3357929 0.3320386 0.3321684
#>  [59,] 0.3357154 0.3319763 0.3323083
#>  [60,] 0.3351695 0.3335421 0.3312885
#>  [61,] 0.3389047 0.3304758 0.3306195
#>  [62,] 0.3352661 0.3315875 0.3331464
#>  [63,] 0.3365629 0.3314529 0.3319842
#>  [64,] 0.3361800 0.3318317 0.3319883
#>  [65,] 0.3124411 0.3072474 0.3803115
#>  [66,] 0.3369500 0.3313953 0.3316546
#>  [67,] 0.3358684 0.3319756 0.3321560
#>  [68,] 0.3255671 0.3526430 0.3217898
#>  [69,] 0.3358306 0.3321966 0.3319728
#>  [70,] 0.3374643 0.3312320 0.3313037
#>  [71,] 0.3368702 0.3314736 0.3316562
#>  [72,] 0.3358586 0.3320195 0.3321219
#>  [73,] 0.3359982 0.3319366 0.3320652
#>  [74,] 0.3359307 0.3318089 0.3322604
#>  [75,] 0.3358110 0.3318997 0.3322893
#>  [76,] 0.2996259 0.3335545 0.3668196
#>  [77,] 0.2995983 0.3342398 0.3661619
#>  [78,] 0.2998078 0.3337832 0.3664089
#>  [79,] 0.3012157 0.3349932 0.3637912
#>  [80,] 0.3022633 0.3297295 0.3680072
#>  [81,] 0.2982095 0.3347233 0.3670672
#>  [82,] 0.3013253 0.3309562 0.3677184
#>  [83,] 0.2998786 0.3336325 0.3664889
#>  [84,] 0.2353924 0.3668283 0.3977793
#>  [85,] 0.2996961 0.3340922 0.3662117
#>  [86,] 0.2994409 0.3341822 0.3663769
#>  [87,] 0.2999251 0.3334214 0.3666535
#>  [88,] 0.2992739 0.3340155 0.3667106
#>  [89,] 0.2993295 0.3342045 0.3664660
#>  [90,] 0.2989822 0.3342607 0.3667571
#>  [91,] 0.3000172 0.3351521 0.3648307
#>  [92,] 0.2990488 0.3338075 0.3671438
#>  [93,] 0.3006016 0.3316912 0.3677073
#>  [94,] 0.3000564 0.3332614 0.3666823
#>  [95,] 0.3001560 0.3335775 0.3662665
#>  [96,] 0.2996068 0.3340868 0.3663064
#>  [97,] 0.2997226 0.3342723 0.3660051
#>  [98,] 0.2998354 0.3339305 0.3662340
#>  [99,] 0.2995881 0.3341504 0.3662615
#> [100,] 0.2994883 0.3339121 0.3665995