Localised multiple kernel k-means

Perform the training step of the localised multiple kernel k-means

lmkkmeans(Km, parameters, verbose = FALSE)

Arguments

Km
kernel matrices

parameters
list of parameters

verbose
Boolean flag. If TRUE, at each iteration the iteration number is printed. Defaults to FALSE.

References

Gonen, M. and Margolin, A.A., 2014. Localized data fusion for kernel k-means clustering with application to cancer biology. In Advances in Neural Information Processing Systems (pp. 1305-1313).

Examples

# Intialise 300 x 300 x 3 array containing M kernel matrices
# representing three different types of similarities
# between 300 data points
km <- array(NA, c(300, 300, 3))
# Load kernel matrices
km[,,1] <- as.matrix(read.csv(system.file("extdata",
"kernel-matrix1.csv", package = "klic"), row.names = 1))
km[,,2] <- as.matrix(read.csv(system.file("extdata",
"kernel-matrix2.csv", package = "klic"), row.names = 1))
km[,,3] <- as.matrix(read.csv(system.file("extdata",
"kernel-matrix3.csv", package = "klic"), row.names = 1))
#initalize the parameters of the algorithm
parameters <- list()
#set the number of clusters
parameters$cluster_count <- 2
# set the number of iterations
parameters$iteration_count <- 10
#perform training
state <- lmkkmeans(km, parameters)
#display the clustering
print(state$clustering)

#>   [1] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [38] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#>  [75] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [112] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 2 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [149] 1 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#> [186] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 1 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#> [223] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#> [260] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#> [297] 2 2 2 2

#display the kernel weights
print(state$Theta)

#>             [,1]      [,2]      [,3]
#>   [1,] 0.3396237 0.3366794 0.3236969
#>   [2,] 0.3362360 0.3368899 0.3268742
#>   [3,] 0.3392490 0.3319174 0.3288336
#>   [4,] 0.3380918 0.3327813 0.3291269
#>   [5,] 0.3386952 0.3361747 0.3251301
#>   [6,] 0.3381573 0.3335843 0.3282584
#>   [7,] 0.3393063 0.3329219 0.3277718
#>   [8,] 0.3408783 0.3323314 0.3267903
#>   [9,] 0.3393728 0.3352669 0.3253603
#>  [10,] 0.3386403 0.3354410 0.3259187
#>  [11,] 0.3350138 0.3335219 0.3314643
#>  [12,] 0.3410074 0.3322249 0.3267676
#>  [13,] 0.3376468 0.3376919 0.3246613
#>  [14,] 0.3369351 0.3362432 0.3268217
#>  [15,] 0.3397539 0.3313093 0.3289368
#>  [16,] 0.3427631 0.3324452 0.3247917
#>  [17,] 0.3351571 0.3382968 0.3265461
#>  [18,] 0.3400451 0.3348751 0.3250798
#>  [19,] 0.3393955 0.3363197 0.3242848
#>  [20,] 0.3389568 0.3331094 0.3279338
#>  [21,] 0.3394041 0.3325867 0.3280092
#>  [22,] 0.3375002 0.3335906 0.3289092
#>  [23,] 0.3396502 0.3309226 0.3294272
#>  [24,] 0.3387772 0.3358047 0.3254181
#>  [25,] 0.3380361 0.3363520 0.3256119
#>  [26,] 0.3389901 0.3341601 0.3268498
#>  [27,] 0.3406318 0.3353725 0.3239956
#>  [28,] 0.3390023 0.3346884 0.3263093
#>  [29,] 0.3395180 0.3359282 0.3245538
#>  [30,] 0.3397662 0.3331998 0.3270340
#>  [31,] 0.3398931 0.3338916 0.3262152
#>  [32,] 0.3381280 0.3366815 0.3251905
#>  [33,] 0.3370987 0.3330210 0.3298803
#>  [34,] 0.3388966 0.3356016 0.3255018
#>  [35,] 0.3397496 0.3335593 0.3266911
#>  [36,] 0.3377013 0.3364261 0.3258726
#>  [37,] 0.3383869 0.3357738 0.3258393
#>  [38,] 0.3377648 0.3371488 0.3250864
#>  [39,] 0.3381349 0.3356140 0.3262511
#>  [40,] 0.3382613 0.3354464 0.3262923
#>  [41,] 0.3390759 0.3364285 0.3244956
#>  [42,] 0.3413436 0.3312584 0.3273979
#>  [43,] 0.3368763 0.3377144 0.3254093
#>  [44,] 0.3383554 0.3349531 0.3266915
#>  [45,] 0.3341216 0.3345946 0.3312837
#>  [46,] 0.3354189 0.3353462 0.3292348
#>  [47,] 0.3404594 0.3321978 0.3273427
#>  [48,] 0.3381074 0.3358356 0.3260570
#>  [49,] 0.3381793 0.3366876 0.3251331
#>  [50,] 0.3369025 0.3371180 0.3259795
#>  [51,] 0.3340036 0.3439437 0.3220528
#>  [52,] 0.3309727 0.3387579 0.3302695
#>  [53,] 0.3276682 0.3458737 0.3264581
#>  [54,] 0.3331555 0.3427314 0.3241131
#>  [55,] 0.3338011 0.3524693 0.3137296
#>  [56,] 0.3336042 0.3444506 0.3219452
#>  [57,] 0.3267704 0.3489908 0.3242388
#>  [58,] 0.3319593 0.3414149 0.3266258
#>  [59,] 0.3318963 0.3425312 0.3255725
#>  [60,] 0.3380833 0.3478817 0.3140350
#>  [61,] 0.3294899 0.3440411 0.3264690
#>  [62,] 0.3250570 0.3447514 0.3301915
#>  [63,] 0.3364874 0.3309942 0.3325184
#>  [64,] 0.3414245 0.3297939 0.3287816
#>  [65,] 0.3308255 0.3379446 0.3312299
#>  [66,] 0.3325680 0.3342607 0.3331713
#>  [67,] 0.3372688 0.3484497 0.3142815
#>  [68,] 0.3348246 0.3404112 0.3247642
#>  [69,] 0.3261689 0.3374736 0.3363575
#>  [70,] 0.3377153 0.3270150 0.3352697
#>  [71,] 0.3354548 0.3289693 0.3355759
#>  [72,] 0.3137277 0.3512194 0.3350528
#>  [73,] 0.3269195 0.3447122 0.3283683
#>  [74,] 0.3380286 0.3270025 0.3349688
#>  [75,] 0.3311376 0.3459069 0.3229556
#>  [76,] 0.3377151 0.3313574 0.3309274
#>  [77,] 0.3328887 0.3527416 0.3143697
#>  [78,] 0.3339476 0.3298247 0.3362277
#>  [79,] 0.3290830 0.3354302 0.3354867
#>  [80,] 0.3321359 0.3429778 0.3248862
#>  [81,] 0.3284123 0.3459518 0.3256359
#>  [82,] 0.3355844 0.3440425 0.3203731
#>  [83,] 0.3306589 0.3465098 0.3228312
#>  [84,] 0.3193248 0.3441378 0.3365374
#>  [85,] 0.3309798 0.3431537 0.3258666
#>  [86,] 0.3325979 0.3409346 0.3264675
#>  [87,] 0.3269599 0.3471251 0.3259150
#>  [88,] 0.3289603 0.3437246 0.3273152
#>  [89,] 0.3294184 0.3445019 0.3260796
#>  [90,] 0.3330948 0.3342693 0.3326360
#>  [91,] 0.3318896 0.3487469 0.3193635
#>  [92,] 0.3263618 0.3464906 0.3271476
#>  [93,] 0.3343407 0.3369629 0.3286964
#>  [94,] 0.3400201 0.3259552 0.3340248
#>  [95,] 0.3366720 0.3253363 0.3379918
#>  [96,] 0.3324107 0.3399705 0.3276187
#>  [97,] 0.3384579 0.3343723 0.3271698
#>  [98,] 0.3259691 0.3452121 0.3288188
#>  [99,] 0.3325904 0.3445803 0.3228293
#> [100,] 0.3239675 0.3489908 0.3270417
#> [101,] 0.3328785 0.3286009 0.3385206
#> [102,] 0.3196662 0.3542283 0.3261055
#> [103,] 0.3450342 0.3210609 0.3339049
#> [104,] 0.3349637 0.3237340 0.3413023
#> [105,] 0.3388151 0.3543872 0.3067977
#> [106,] 0.3114552 0.3578128 0.3307320
#> [107,] 0.3221180 0.3493011 0.3285808
#> [108,] 0.3460872 0.3084389 0.3454739
#> [109,] 0.3456702 0.3193194 0.3350103
#> [110,] 0.3162569 0.3436616 0.3400814
#> [111,] 0.3191887 0.3578452 0.3229661
#> [112,] 0.3347811 0.3264167 0.3388022
#> [113,] 0.3352358 0.3259796 0.3387846
#> [114,] 0.3355407 0.3276277 0.3368316
#> [115,] 0.3479237 0.3171459 0.3349304
#> [116,] 0.3130415 0.3573747 0.3295839
#> [117,] 0.3361364 0.3468569 0.3170066
#> [118,] 0.3228747 0.3495253 0.3276000
#> [119,] 0.3199017 0.3543549 0.3257434
#> [120,] 0.3107057 0.3555470 0.3337473
#> [121,] 0.3210896 0.3541521 0.3247583
#> [122,] 0.3169443 0.3525047 0.3305510
#> [123,] 0.3329187 0.3301699 0.3369115
#> [124,] 0.3358679 0.3267752 0.3373569
#> [125,] 0.3227136 0.3492991 0.3279873
#> [126,] 0.3194826 0.3546722 0.3258452
#> [127,] 0.3181013 0.3564679 0.3254308
#> [128,] 0.3343232 0.3339627 0.3317141
#> [129,] 0.3275084 0.3290460 0.3434456
#> [130,] 0.3367455 0.3437968 0.3194577
#> [131,] 0.3344770 0.3268848 0.3386383
#> [132,] 0.3288997 0.3511457 0.3199546
#> [133,] 0.3221303 0.3529499 0.3249199
#> [134,] 0.3130010 0.3545084 0.3324906
#> [135,] 0.3301983 0.3328629 0.3369388
#> [136,] 0.3318386 0.3339765 0.3341849
#> [137,] 0.3316088 0.3195393 0.3488518
#> [138,] 0.3445593 0.3202934 0.3351474
#> [139,] 0.3343272 0.3295962 0.3360766
#> [140,] 0.3352814 0.3273710 0.3373475
#> [141,] 0.3281272 0.3300715 0.3418014
#> [142,] 0.3275852 0.3542600 0.3181548
#> [143,] 0.3318369 0.3303094 0.3378537
#> [144,] 0.3317182 0.3288162 0.3394656
#> [145,] 0.3242240 0.3493163 0.3264597
#> [146,] 0.3198525 0.3509948 0.3291527
#> [147,] 0.3167791 0.3482192 0.3350017
#> [148,] 0.3281571 0.3497531 0.3220899
#> [149,] 0.3354503 0.3267217 0.3378280
#> [150,] 0.3128058 0.3577620 0.3294322
#> [151,] 0.3349399 0.3299030 0.3351571
#> [152,] 0.3425939 0.3261785 0.3312276
#> [153,] 0.3216558 0.3318310 0.3465131
#> [154,] 0.3399879 0.3171727 0.3428394
#> [155,] 0.3401379 0.3173196 0.3425425
#> [156,] 0.3379595 0.3196850 0.3423555
#> [157,] 0.3401342 0.3190831 0.3407828
#> [158,] 0.3334473 0.3313466 0.3352061
#> [159,] 0.3405591 0.3394084 0.3200325
#> [160,] 0.3357258 0.3303345 0.3339398
#> [161,] 0.3418917 0.3371538 0.3209545
#> [162,] 0.3220304 0.3314347 0.3465350
#> [163,] 0.3251629 0.3304649 0.3443722
#> [164,] 0.3398129 0.3197456 0.3404416
#> [165,] 0.3396615 0.3175821 0.3427564
#> [166,] 0.3380102 0.3240566 0.3379332
#> [167,] 0.3403861 0.3168308 0.3427831
#> [168,] 0.3398281 0.3199076 0.3402642
#> [169,] 0.3364512 0.3290376 0.3345112
#> [170,] 0.3343020 0.3308337 0.3348642
#> [171,] 0.3414179 0.3357379 0.3228441
#> [172,] 0.3416198 0.3155448 0.3428354
#> [173,] 0.3355909 0.3325175 0.3318916
#> [174,] 0.3398327 0.3190667 0.3411006
#> [175,] 0.3401426 0.3173397 0.3425177
#> [176,] 0.3217186 0.3308125 0.3474689
#> [177,] 0.3424685 0.3158950 0.3416365
#> [178,] 0.3408761 0.3170853 0.3420386
#> [179,] 0.3360544 0.3287276 0.3352181
#> [180,] 0.3423539 0.3221938 0.3354523
#> [181,] 0.3285719 0.3361167 0.3353114
#> [182,] 0.3444902 0.3268628 0.3286470
#> [183,] 0.3407319 0.3220506 0.3372175
#> [184,] 0.3320431 0.3339675 0.3339893
#> [185,] 0.3434438 0.3227073 0.3338489
#> [186,] 0.3436078 0.3232207 0.3331714
#> [187,] 0.3397538 0.3189628 0.3412834
#> [188,] 0.3400589 0.3192841 0.3406571
#> [189,] 0.3362104 0.3274055 0.3363842
#> [190,] 0.3429725 0.3272043 0.3298232
#> [191,] 0.3395131 0.3180508 0.3424361
#> [192,] 0.3413411 0.3364564 0.3222025
#> [193,] 0.3380744 0.3251538 0.3367718
#> [194,] 0.3334751 0.3302360 0.3362889
#> [195,] 0.3404510 0.3187120 0.3408369
#> [196,] 0.3405095 0.3182954 0.3411951
#> [197,] 0.3414598 0.3158705 0.3426696
#> [198,] 0.3408678 0.3173841 0.3417482
#> [199,] 0.3414220 0.3160475 0.3425305
#> [200,] 0.3297730 0.3329438 0.3372832
#> [201,] 0.3370876 0.3342888 0.3286236
#> [202,] 0.3341808 0.3311220 0.3346972
#> [203,] 0.3362156 0.3264632 0.3373212
#> [204,] 0.3365781 0.3332035 0.3302185
#> [205,] 0.3310770 0.3340459 0.3348770
#> [206,] 0.3332612 0.3328474 0.3338915
#> [207,] 0.3320631 0.3295761 0.3383608
#> [208,] 0.3328847 0.3283480 0.3387673
#> [209,] 0.3321314 0.3297152 0.3381535
#> [210,] 0.3305317 0.3312690 0.3381993
#> [211,] 0.3327664 0.3209981 0.3462355
#> [212,] 0.3290490 0.3283768 0.3425741
#> [213,] 0.3316062 0.3324866 0.3359072
#> [214,] 0.3341117 0.3333258 0.3325625
#> [215,] 0.3415385 0.3281102 0.3303513
#> [216,] 0.3412715 0.3299969 0.3287316
#> [217,] 0.3408291 0.3292840 0.3298869
#> [218,] 0.3359765 0.3285177 0.3355058
#> [219,] 0.3351574 0.3292317 0.3356109
#> [220,] 0.3379507 0.3244401 0.3376092
#> [221,] 0.3384795 0.3258244 0.3356960
#> [222,] 0.3393388 0.3330154 0.3276458
#> [223,] 0.3343937 0.3304721 0.3351342
#> [224,] 0.3392049 0.3293965 0.3313986
#> [225,] 0.3334442 0.3252306 0.3413252
#> [226,] 0.3388878 0.3301649 0.3309473
#> [227,] 0.3277384 0.3279519 0.3443097
#> [228,] 0.3391068 0.3249402 0.3359530
#> [229,] 0.3368705 0.3351168 0.3280127
#> [230,] 0.3396087 0.3250636 0.3353277
#> [231,] 0.3375996 0.3306533 0.3317471
#> [232,] 0.3384168 0.3238462 0.3377370
#> [233,] 0.3366613 0.3343378 0.3290009
#> [234,] 0.3326468 0.3308123 0.3365409
#> [235,] 0.3377980 0.3271769 0.3350251
#> [236,] 0.3373753 0.3310851 0.3315397
#> [237,] 0.3374840 0.3335447 0.3289714
#> [238,] 0.3358965 0.3353939 0.3287097
#> [239,] 0.3392218 0.3294320 0.3313462
#> [240,] 0.3353755 0.3355351 0.3290894
#> [241,] 0.3368378 0.3297595 0.3334027
#> [242,] 0.3438566 0.3275780 0.3285653
#> [243,] 0.3382184 0.3255083 0.3362734
#> [244,] 0.3346564 0.3356967 0.3296470
#> [245,] 0.3380896 0.3327061 0.3292042
#> [246,] 0.3393605 0.3298151 0.3308244
#> [247,] 0.3393577 0.3347708 0.3258715
#> [248,] 0.3307472 0.3306701 0.3385827
#> [249,] 0.3342365 0.3296244 0.3361391
#> [250,] 0.3354039 0.3345806 0.3300155
#> [251,] 0.3358100 0.3228594 0.3413306
#> [252,] 0.3328261 0.3290004 0.3381735
#> [253,] 0.3328250 0.3263957 0.3407793
#> [254,] 0.3320621 0.3261722 0.3417657
#> [255,] 0.3340926 0.3255805 0.3403269
#> [256,] 0.3334816 0.3258011 0.3407173
#> [257,] 0.3331802 0.3255950 0.3412248
#> [258,] 0.3352391 0.3204921 0.3442688
#> [259,] 0.3303509 0.3250367 0.3446124
#> [260,] 0.3336186 0.3278508 0.3385306
#> [261,] 0.3341105 0.3272823 0.3386073
#> [262,] 0.3327606 0.3276027 0.3396367
#> [263,] 0.3314504 0.3269620 0.3415876
#> [264,] 0.3329308 0.3268445 0.3402247
#> [265,] 0.3354142 0.3260389 0.3385469
#> [266,] 0.3388023 0.3369157 0.3242820
#> [267,] 0.3304922 0.3274678 0.3420400
#> [268,] 0.3342259 0.3269363 0.3388378
#> [269,] 0.3345229 0.3270465 0.3384306
#> [270,] 0.3328298 0.3272265 0.3399437
#> [271,] 0.3353757 0.3199807 0.3446437
#> [272,] 0.3341530 0.3251046 0.3407424
#> [273,] 0.3281547 0.3286155 0.3432297
#> [274,] 0.3328022 0.3279205 0.3392773
#> [275,] 0.3333524 0.3284329 0.3382147
#> [276,] 0.3363620 0.3311886 0.3324495
#> [277,] 0.3311188 0.3289080 0.3399732
#> [278,] 0.3337815 0.3255781 0.3406404
#> [279,] 0.3350201 0.3220795 0.3429005
#> [280,] 0.3340787 0.3194730 0.3464483
#> [281,] 0.3316840 0.3270812 0.3412348
#> [282,] 0.3373100 0.3300065 0.3326835
#> [283,] 0.3294164 0.3261119 0.3444717
#> [284,] 0.3369826 0.3311724 0.3318451
#> [285,] 0.3329448 0.3249763 0.3420789
#> [286,] 0.3312984 0.3253437 0.3433579
#> [287,] 0.3336635 0.3209901 0.3453463
#> [288,] 0.3336842 0.3253712 0.3409445
#> [289,] 0.3312529 0.3276636 0.3410834
#> [290,] 0.3300817 0.3265467 0.3433715
#> [291,] 0.3343547 0.3190942 0.3465511
#> [292,] 0.3362025 0.3222164 0.3415811
#> [293,] 0.3309554 0.3276623 0.3413822
#> [294,] 0.3348880 0.3231952 0.3419168
#> [295,] 0.3324663 0.3264415 0.3410923
#> [296,] 0.3333891 0.3274092 0.3392017
#> [297,] 0.3295168 0.3285759 0.3419073
#> [298,] 0.3317622 0.3270624 0.3411755
#> [299,] 0.3323253 0.3268466 0.3408281
#> [300,] 0.3337041 0.3276929 0.3386030