Localised multiple kernel k-means

Perform the training step of the localised multiple kernel k-means

lmkkmeans_missingData(Km, parameters, missing = NULL, verbose = FALSE)

Arguments

Km
Array of size N X N X M containing M kernel matrices.

parameters
List of parameters, containing "iteration_count", "cluster_count".

missing
Matrix of size N X M containing missingness indicators, i.e. missing[i,j] = 1 (or = TRUE) if observation "i" is missing in dataset "j", missing[i,j] = 0 (or = FALSE).

verbose
Boolean flag. If TRUE, at each iteration the iteration number is printed. Defaults to FALSE.

References

Gonen, M. and Margolin, A.A., 2014. Localized data fusion for kernel k-means clustering with application to cancer biology. In Advances in Neural Information Processing Systems (pp. 1305-1313).

Examples

# Intialise 100 x 100 x 3 array containing M kernel matrices
# representing three different types of similarities
# between 100 data points
km <- array(NA, c(100, 100, 3))
# Load kernel matrices
km[,,1] <- as.matrix(read.csv(system.file('extdata',
'kernel_matrix1.csv', package = 'klic'), row.names = 1))
km[,,2] <- as.matrix(read.csv(system.file('extdata',
'kernel_matrix2.csv', package = 'klic'), row.names = 1))
km[,,3] <- as.matrix(read.csv(system.file('extdata',
'kernel_matrix3.csv', package = 'klic'), row.names = 1))
# Introduce some missing data
km[76:80, , 1] <- NA
km[, 76:80, 1] <- NA

# Define missingness indicators
missing <- matrix(FALSE, 100, 3)
missing[76:80,1] <- TRUE

# Initalize the parameters of the algorithm
parameters <- list()
# Set the number of clusters
parameters$cluster_count <- 2
# Set the number of iterations
parameters$iteration_count <- 10

# Perform training
state <- lmkkmeans_missingData(km, parameters, missing)

# Display the clustering
print(state$clustering)

#>   [1] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [38] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 1 2 2 2 2 2 2 2 2 2
#>  [75] 2 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1

# Display the kernel weights
print(state$Theta)

#>             [,1]      [,2]      [,3]
#>   [1,] 0.3345422 0.3309156 0.3345422
#>   [2,] 0.3346034 0.3307932 0.3346034
#>   [3,] 0.3332602 0.3334796 0.3332602
#>   [4,] 0.3343382 0.3313235 0.3343382
#>   [5,] 0.3341747 0.3316506 0.3341747
#>   [6,] 0.3340609 0.3318782 0.3340609
#>   [7,] 0.3337827 0.3324347 0.3337827
#>   [8,] 0.3340189 0.3319623 0.3340189
#>   [9,] 0.3331686 0.3336628 0.3331686
#>  [10,] 0.3341359 0.3317283 0.3341359
#>  [11,] 0.3343161 0.3313678 0.3343161
#>  [12,] 0.3340752 0.3318497 0.3340752
#>  [13,] 0.3341190 0.3317621 0.3341190
#>  [14,] 0.3343812 0.3312376 0.3343812
#>  [15,] 0.3263283 0.3473434 0.3263283
#>  [16,] 0.3341919 0.3316162 0.3341919
#>  [17,] 0.3332004 0.3335993 0.3332004
#>  [18,] 0.3343645 0.3312711 0.3343645
#>  [19,] 0.3344333 0.3311333 0.3344333
#>  [20,] 0.3344728 0.3310545 0.3344728
#>  [21,] 0.2856434 0.4287131 0.2856434
#>  [22,] 0.3341323 0.3317353 0.3341323
#>  [23,] 0.2874358 0.4251285 0.2874358
#>  [24,] 0.3335968 0.3328065 0.3335968
#>  [25,] 0.3342161 0.3315678 0.3342161
#>  [26,] 0.3045392 0.3812319 0.3142289
#>  [27,] 0.3030332 0.3842848 0.3126820
#>  [28,] 0.3030612 0.3846001 0.3123387
#>  [29,] 0.3020791 0.3861326 0.3117883
#>  [30,] 0.3028855 0.3844765 0.3126380
#>  [31,] 0.3041698 0.3821638 0.3136664
#>  [32,] 0.3028975 0.3849564 0.3121461
#>  [33,] 0.3023410 0.3854863 0.3121727
#>  [34,] 0.3031912 0.3843219 0.3124870
#>  [35,] 0.3028431 0.3848100 0.3123468
#>  [36,] 0.3021412 0.3859056 0.3119533
#>  [37,] 0.3034657 0.3836329 0.3129014
#>  [38,] 0.3034096 0.3840920 0.3124984
#>  [39,] 0.3030047 0.3843508 0.3126445
#>  [40,] 0.3021813 0.3857974 0.3120213
#>  [41,] 0.3023228 0.3855213 0.3121559
#>  [42,] 0.3048188 0.3812570 0.3139242
#>  [43,] 0.3029699 0.3846735 0.3123566
#>  [44,] 0.3128003 0.3652740 0.3219256
#>  [45,] 0.3030333 0.3843654 0.3126013
#>  [46,] 0.3088977 0.3731551 0.3179471
#>  [47,] 0.2831689 0.3940957 0.3227354
#>  [48,] 0.3035839 0.3833754 0.3130407
#>  [49,] 0.3023094 0.3856619 0.3120287
#>  [50,] 0.3038377 0.3830581 0.3131043
#>  [51,] 0.3339867 0.3327837 0.3332296
#>  [52,] 0.3340490 0.3327033 0.3332477
#>  [53,] 0.3339912 0.3327959 0.3332129
#>  [54,] 0.3340969 0.3325723 0.3333308
#>  [55,] 0.3339968 0.3327261 0.3332771
#>  [56,] 0.2819594 0.4377925 0.2802481
#>  [57,] 0.3340157 0.3327638 0.3332204
#>  [58,] 0.3340239 0.3327141 0.3332620
#>  [59,] 0.3340875 0.3325637 0.3333489
#>  [60,] 0.3339288 0.3330114 0.3330598
#>  [61,] 0.3333851 0.3330564 0.3335586
#>  [62,] 0.3338513 0.3326618 0.3334869
#>  [63,] 0.3341439 0.3324855 0.3333706
#>  [64,] 0.3339070 0.3327970 0.3332960
#>  [65,] 0.3034295 0.3022743 0.3942962
#>  [66,] 0.3339769 0.3326485 0.3333746
#>  [67,] 0.3343843 0.3325272 0.3330885
#>  [68,] 0.3199483 0.3612887 0.3187630
#>  [69,] 0.3340056 0.3327360 0.3332585
#>  [70,] 0.3339034 0.3327813 0.3333153
#>  [71,] 0.3338886 0.3328173 0.3332941
#>  [72,] 0.3340325 0.3327011 0.3332664
#>  [73,] 0.3340765 0.3326914 0.3332321
#>  [74,] 0.3340836 0.3326055 0.3333108
#>  [75,] 0.3340161 0.3326582 0.3333257
#>  [76,] 0.0000000 0.4649590 0.5350410
#>  [77,] 0.0000000 0.4659715 0.5340285
#>  [78,] 0.0000000 0.4655706 0.5344294
#>  [79,] 0.0000000 0.4702838 0.5297162
#>  [80,] 0.0000000 0.4583555 0.5416445
#>  [81,] 0.1148151 0.4114580 0.4737270
#>  [82,] 0.1227397 0.4029247 0.4743356
#>  [83,] 0.1160982 0.4105192 0.4733826
#>  [84,] 0.0835433 0.4244873 0.4919694
#>  [85,] 0.1158499 0.4111683 0.4729818
#>  [86,] 0.1154082 0.4113922 0.4731995
#>  [87,] 0.1163661 0.4102659 0.4733680
#>  [88,] 0.1152958 0.4110802 0.4736240
#>  [89,] 0.1153411 0.4113957 0.4732632
#>  [90,] 0.1152217 0.4111723 0.4736060
#>  [91,] 0.1184362 0.4123554 0.4692083
#>  [92,] 0.1148473 0.4108919 0.4742607
#>  [93,] 0.1212532 0.4044454 0.4743014
#>  [94,] 0.1160927 0.4102608 0.4736466
#>  [95,] 0.1162248 0.4106382 0.4731371
#>  [96,] 0.1157666 0.4111465 0.4730869
#>  [97,] 0.1155752 0.4115133 0.4729114
#>  [98,] 0.1155980 0.4112035 0.4731985
#>  [99,] 0.1152405 0.4115440 0.4732155
#> [100,] 0.1156212 0.4110076 0.4733711