Localised multiple kernel k-means

Perform the training step of the localised multiple kernel k-means

lmkkmeans_missingData(Km, parameters, missing = NULL, verbose = FALSE)

Arguments

Km
Array of size N X N X M containing M kernel matrices.

parameters
List of parameters, containing `iteration_count`, `cluster_count`.

missing
Matrix of size N X M containing missingness indicators, i.e. missing[i,j] = 1 (or =TRUE) if observation `i` is missing in dataset `j`, missing[i,j]=0 (or =FALSE).

verbose
Boolean flag. If TRUE, at each iteration the iteration number is printed. Defaults to FALSE.

References

Gonen, M. and Margolin, A.A., 2014. Localized data fusion for kernel k-means clustering with application to cancer biology. In Advances in Neural Information Processing Systems (pp. 1305-1313).

Examples

# Intialise 300 x 300 x 3 array containing M kernel matrices
# representing three different types of similarities
# between 300 data points
km <- array(NA, c(300, 300, 3))
# Load kernel matrices
km[,,1] <- as.matrix(read.csv(system.file("extdata",
                                         "kernel-matrix1.csv", package = "klic"), row.names = 1))
km[,,2] <- as.matrix(read.csv(system.file("extdata",
                                         "kernel-matrix2.csv", package = "klic"), row.names = 1))
km[,,3] <- as.matrix(read.csv(system.file("extdata",
                                         "kernel-matrix3.csv", package = "klic"), row.names = 1))
# Introduce some missing data
km[76:100, , 1] <- NA
km[, 76:100, 1] <- NA

# Define missingness indicators
missing <- matrix(FALSE, 300, 3)
missing[76:100,1] <- TRUE

#initalize the parameters of the algorithm
parameters <- list()
#set the number of clusters
parameters$cluster_count <- 2
# set the number of iterations
parameters$iteration_count <- 10
#perform training
state <- lmkkmeans_missingData(km, parameters, missing)
#display the clustering
print(state$clustering)

#>   [1] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [38] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#>  [75] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2
#> [112] 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 2 1 2 2 2 2 2 1 2 2 2 2 2 2 2 2 2 2 2 2 2
#> [149] 2 2 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [186] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [223] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [260] 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1 1
#> [297] 1 1 1 1

#display the kernel weights
print(state$Theta)

#>             [,1]      [,2]      [,3]
#>   [1,] 0.3306851 0.3373085 0.3320064
#>   [2,] 0.3303302 0.3370176 0.3326522
#>   [3,] 0.3286797 0.3319145 0.3394058
#>   [4,] 0.3180668 0.3234631 0.3584701
#>   [5,] 0.3245003 0.3343385 0.3411613
#>   [6,] 0.3246284 0.3488152 0.3265564
#>   [7,] 0.3319502 0.3353126 0.3327372
#>   [8,] 0.3315064 0.3341717 0.3343219
#>   [9,] 0.3310706 0.3379779 0.3309515
#>  [10,] 0.3308094 0.3371288 0.3320618
#>  [11,] 0.3116433 0.3208592 0.3674975
#>  [12,] 0.3313942 0.3337672 0.3348386
#>  [13,] 0.3308418 0.3386037 0.3305545
#>  [14,] 0.3302678 0.3373605 0.3323716
#>  [15,] 0.3320283 0.3345125 0.3334592
#>  [16,] 0.3331004 0.3349586 0.3319411
#>  [17,] 0.3284418 0.3369247 0.3346335
#>  [18,] 0.3285373 0.3398253 0.3316374
#>  [19,] 0.3313203 0.3372037 0.3314760
#>  [20,] 0.3316026 0.3354634 0.3329340
#>  [21,] 0.3293118 0.3405821 0.3301061
#>  [22,] 0.3305267 0.3347273 0.3347460
#>  [23,] 0.3320915 0.3343939 0.3335146
#>  [24,] 0.3310376 0.3371931 0.3317692
#>  [25,] 0.3309233 0.3374430 0.3316337
#>  [26,] 0.3312044 0.3361629 0.3326328
#>  [27,] 0.3319973 0.3365908 0.3314119
#>  [28,] 0.3236994 0.3517586 0.3245420
#>  [29,] 0.3309702 0.3377263 0.3313035
#>  [30,] 0.3319406 0.3355413 0.3325182
#>  [31,] 0.3295307 0.3407274 0.3297419
#>  [32,] 0.3308164 0.3378405 0.3313430
#>  [33,] 0.3128799 0.3209702 0.3661499
#>  [34,] 0.3312552 0.3373093 0.3314354
#>  [35,] 0.3296786 0.3402768 0.3300446
#>  [36,] 0.3304052 0.3382959 0.3312988
#>  [37,] 0.3308848 0.3374585 0.3316567
#>  [38,] 0.3309150 0.3378059 0.3312790
#>  [39,] 0.3295992 0.3365635 0.3338374
#>  [40,] 0.3300136 0.3387958 0.3311906
#>  [41,] 0.3311761 0.3376710 0.3311529
#>  [42,] 0.3323634 0.3347772 0.3328594
#>  [43,] 0.3295136 0.3393125 0.3311739
#>  [44,] 0.3296743 0.3393355 0.3309902
#>  [45,] 0.3131125 0.3229406 0.3639469
#>  [46,] 0.3305285 0.3362220 0.3332496
#>  [47,] 0.3301622 0.3361961 0.3336417
#>  [48,] 0.3307482 0.3372071 0.3320447
#>  [49,] 0.3309728 0.3375579 0.3314693
#>  [50,] 0.3296099 0.3371216 0.3332685
#>  [51,] 0.2526705 0.4057735 0.3415559
#>  [52,] 0.2518080 0.3880302 0.3601618
#>  [53,] 0.2331879 0.4109689 0.3558432
#>  [54,] 0.2618101 0.3989035 0.3392864
#>  [55,] 0.2870350 0.4232368 0.2897282
#>  [56,] 0.2551850 0.4046683 0.3401466
#>  [57,] 0.2493356 0.4126443 0.3380201
#>  [58,] 0.2379253 0.4028961 0.3591786
#>  [59,] 0.2553636 0.3983318 0.3463046
#>  [60,] 0.2841842 0.4173184 0.2984974
#>  [61,] 0.2365308 0.4063154 0.3571538
#>  [62,] 0.2345634 0.4047484 0.3606882
#>  [63,] 0.2646183 0.3649545 0.3704271
#>  [64,] 0.2999018 0.3390702 0.3610280
#>  [65,] 0.2602192 0.3756162 0.3641645
#>  [66,] 0.2608887 0.3699219 0.3691894
#>  [67,] 0.2742525 0.4168007 0.3089467
#>  [68,] 0.2535554 0.3947374 0.3517072
#>  [69,] 0.2484938 0.3838739 0.3676323
#>  [70,] 0.3000607 0.3157557 0.3841836
#>  [71,] 0.3019596 0.3144755 0.3835649
#>  [72,] 0.2505023 0.3947945 0.3547032
#>  [73,] 0.2378579 0.4040826 0.3580595
#>  [74,] 0.3002194 0.3197281 0.3800524
#>  [75,] 0.2613124 0.4004777 0.3382099
#>  [76,] 0.0000000 0.4657503 0.5342497
#>  [77,] 0.0000000 0.5604572 0.4395428
#>  [78,] 0.0000000 0.4647839 0.5352161
#>  [79,] 0.0000000 0.5053835 0.4946165
#>  [80,] 0.0000000 0.5281095 0.4718905
#>  [81,] 0.0000000 0.5349154 0.4650846
#>  [82,] 0.0000000 0.5419808 0.4580192
#>  [83,] 0.0000000 0.5350605 0.4649395
#>  [84,] 0.0000000 0.5201924 0.4798076
#>  [85,] 0.0000000 0.5289486 0.4710514
#>  [86,] 0.0000000 0.5260602 0.4739398
#>  [87,] 0.0000000 0.5355681 0.4644319
#>  [88,] 0.0000000 0.5272467 0.4727533
#>  [89,] 0.0000000 0.5292334 0.4707666
#>  [90,] 0.0000000 0.5068199 0.4931801
#>  [91,] 0.0000000 0.5546680 0.4453320
#>  [92,] 0.0000000 0.5310783 0.4689217
#>  [93,] 0.0000000 0.5145751 0.4854249
#>  [94,] 0.0000000 0.4789085 0.5210915
#>  [95,] 0.0000000 0.4599935 0.5400065
#>  [96,] 0.0000000 0.5238254 0.4761746
#>  [97,] 0.0000000 0.5053568 0.4946432
#>  [98,] 0.0000000 0.5284616 0.4715384
#>  [99,] 0.0000000 0.5352421 0.4647579
#> [100,] 0.0000000 0.5344341 0.4655659
#> [101,] 0.3266237 0.3285269 0.3448495
#> [102,] 0.2823775 0.3999131 0.3177093
#> [103,] 0.3167583 0.3326940 0.3505476
#> [104,] 0.3270791 0.3179553 0.3549657
#> [105,] 0.2921716 0.4085992 0.2992291
#> [106,] 0.2715493 0.4114976 0.3169530
#> [107,] 0.2949542 0.3821743 0.3228715
#> [108,] 0.3130611 0.3089247 0.3780143
#> [109,] 0.3237051 0.3283608 0.3479341
#> [110,] 0.2681172 0.3683882 0.3634946
#> [111,] 0.2788330 0.4141497 0.3070173
#> [112,] 0.3295996 0.3248205 0.3455799
#> [113,] 0.3279789 0.3218779 0.3501432
#> [114,] 0.3298416 0.3278317 0.3423268
#> [115,] 0.3217636 0.3265194 0.3517170
#> [116,] 0.2785316 0.4057773 0.3156911
#> [117,] 0.2894056 0.3910566 0.3195378
#> [118,] 0.2960445 0.3830312 0.3209243
#> [119,] 0.2840183 0.4007050 0.3152767
#> [120,] 0.2696158 0.3940728 0.3363114
#> [121,] 0.2879539 0.3981888 0.3138572
#> [122,] 0.2908413 0.3847133 0.3244454
#> [123,] 0.3242787 0.3331063 0.3426151
#> [124,] 0.3295566 0.3264074 0.3440360
#> [125,] 0.2941246 0.3839496 0.3219257
#> [126,] 0.2856397 0.3996945 0.3146658
#> [127,] 0.2791206 0.4085215 0.3123579
#> [128,] 0.2682121 0.3669517 0.3648361
#> [129,] 0.3230745 0.3305056 0.3464199
#> [130,] 0.2871951 0.3858013 0.3270035
#> [131,] 0.3292669 0.3254740 0.3452591
#> [132,] 0.3117394 0.3859593 0.3023013
#> [133,] 0.2928970 0.3929663 0.3141367
#> [134,] 0.2564190 0.4044776 0.3391034
#> [135,] 0.3295952 0.3329748 0.3374300
#> [136,] 0.3194668 0.3436131 0.3369201
#> [137,] 0.3155279 0.3042288 0.3802433
#> [138,] 0.3224066 0.3294156 0.3481778
#> [139,] 0.3262606 0.3324845 0.3412549
#> [140,] 0.3304041 0.3269966 0.3425994
#> [141,] 0.3194533 0.3275441 0.3530027
#> [142,] 0.3060955 0.4022994 0.2916052
#> [143,] 0.3198840 0.3333809 0.3467351
#> [144,] 0.3282604 0.3260002 0.3457394
#> [145,] 0.2958684 0.3830891 0.3210425
#> [146,] 0.2839365 0.3898194 0.3262442
#> [147,] 0.2745496 0.3788130 0.3466374
#> [148,] 0.3082697 0.3805545 0.3111757
#> [149,] 0.3289609 0.3249420 0.3460971
#> [150,] 0.2725852 0.4121630 0.3152518
#> [151,] 0.3341349 0.3316174 0.3342477
#> [152,] 0.3437093 0.3254787 0.3308119
#> [153,] 0.3196024 0.3335731 0.3468245
#> [154,] 0.3408444 0.3154742 0.3436814
#> [155,] 0.3413404 0.3153380 0.3433216
#> [156,] 0.3384867 0.3187657 0.3427475
#> [157,] 0.3411096 0.3178520 0.3410384
#> [158,] 0.3327697 0.3332752 0.3339550
#> [159,] 0.3399866 0.3425809 0.3174325
#> [160,] 0.3361372 0.3314772 0.3323855
#> [161,] 0.3416612 0.3399292 0.3184096
#> [162,] 0.3196840 0.3319961 0.3483199
#> [163,] 0.3236495 0.3320578 0.3442927
#> [164,] 0.3406113 0.3185960 0.3407927
#> [165,] 0.3406971 0.3157859 0.3435170
#> [166,] 0.3379108 0.3248035 0.3372857
#> [167,] 0.3413231 0.3151421 0.3435349
#> [168,] 0.3407120 0.3188166 0.3404714
#> [169,] 0.3370401 0.3299722 0.3329877
#> [170,] 0.3333508 0.3329457 0.3337034
#> [171,] 0.3411663 0.3382699 0.3205638
#> [172,] 0.3427603 0.3136726 0.3435671
#> [173,] 0.3344986 0.3339202 0.3315812
#> [174,] 0.3403675 0.3180267 0.3416058
#> [175,] 0.3412207 0.3156040 0.3431754
#> [176,] 0.3202985 0.3309575 0.3487440
#> [177,] 0.3435218 0.3140870 0.3423912
#> [178,] 0.3419478 0.3153144 0.3427378
#> [179,] 0.3362904 0.3298279 0.3338817
#> [180,] 0.3433610 0.3213123 0.3353267
#> [181,] 0.3263212 0.3384564 0.3352224
#> [182,] 0.3460883 0.3262560 0.3276557
#> [183,] 0.3418672 0.3210950 0.3370379
#> [184,] 0.3327538 0.3348954 0.3323508
#> [185,] 0.3446104 0.3218645 0.3335251
#> [186,] 0.3449399 0.3220176 0.3330425
#> [187,] 0.3409644 0.3174941 0.3415415
#> [188,] 0.3405551 0.3187363 0.3407086
#> [189,] 0.3351313 0.3297567 0.3351121
#> [190,] 0.3443216 0.3266790 0.3289994
#> [191,] 0.3402942 0.3164783 0.3432275
#> [192,] 0.3410139 0.3390396 0.3199465
#> [193,] 0.3379073 0.3261375 0.3359553
#> [194,] 0.3325593 0.3322538 0.3351869
#> [195,] 0.3414315 0.3171539 0.3414145
#> [196,] 0.3413173 0.3166343 0.3420484
#> [197,] 0.3423915 0.3140364 0.3435721
#> [198,] 0.3418386 0.3158526 0.3423088
#> [199,] 0.3425693 0.3142073 0.3432235
#> [200,] 0.3292643 0.3328275 0.3379082
#> [201,] 0.3369133 0.3345702 0.3285165
#> [202,] 0.3346523 0.3320938 0.3332538
#> [203,] 0.3363739 0.3265004 0.3371257
#> [204,] 0.3362055 0.3334689 0.3303256
#> [205,] 0.3318914 0.3332369 0.3348717
#> [206,] 0.3322254 0.3331209 0.3346537
#> [207,] 0.3321281 0.3296656 0.3382063
#> [208,] 0.3329304 0.3281730 0.3388966
#> [209,] 0.3314917 0.3300032 0.3385051
#> [210,] 0.3305089 0.3314446 0.3380464
#> [211,] 0.3335799 0.3208118 0.3456083
#> [212,] 0.3290023 0.3286355 0.3423621
#> [213,] 0.3315403 0.3326441 0.3358155
#> [214,] 0.3348951 0.3343970 0.3307078
#> [215,] 0.3414181 0.3282357 0.3303462
#> [216,] 0.3412673 0.3299739 0.3287589
#> [217,] 0.3400590 0.3300045 0.3299365
#> [218,] 0.3358554 0.3289018 0.3352428
#> [219,] 0.3346333 0.3290723 0.3362944
#> [220,] 0.3377897 0.3248002 0.3374101
#> [221,] 0.3384292 0.3260065 0.3355643
#> [222,] 0.3393723 0.3331782 0.3274495
#> [223,] 0.3347809 0.3311288 0.3340903
#> [224,] 0.3399228 0.3297283 0.3303489
#> [225,] 0.3333833 0.3260975 0.3405192
#> [226,] 0.3391272 0.3297426 0.3311302
#> [227,] 0.3278647 0.3282474 0.3438879
#> [228,] 0.3388697 0.3252683 0.3358621
#> [229,] 0.3357687 0.3358184 0.3284129
#> [230,] 0.3395525 0.3252370 0.3352105
#> [231,] 0.3377405 0.3305397 0.3317198
#> [232,] 0.3384254 0.3240470 0.3375276
#> [233,] 0.3366292 0.3345397 0.3288311
#> [234,] 0.3326135 0.3309483 0.3364382
#> [235,] 0.3370030 0.3281135 0.3348835
#> [236,] 0.3375681 0.3308992 0.3315327
#> [237,] 0.3375572 0.3336152 0.3288276
#> [238,] 0.3359018 0.3356248 0.3284734
#> [239,] 0.3390566 0.3296785 0.3312649
#> [240,] 0.3353960 0.3357186 0.3288854
#> [241,] 0.3375871 0.3302106 0.3322023
#> [242,] 0.3432632 0.3279170 0.3288198
#> [243,] 0.3379194 0.3271902 0.3348904
#> [244,] 0.3344724 0.3359846 0.3295430
#> [245,] 0.3383071 0.3320384 0.3296544
#> [246,] 0.3387268 0.3303110 0.3309621
#> [247,] 0.3393458 0.3349826 0.3256716
#> [248,] 0.3304151 0.3312529 0.3383320
#> [249,] 0.3341687 0.3299071 0.3359242
#> [250,] 0.3356050 0.3343573 0.3300377
#> [251,] 0.3358028 0.3230549 0.3411423
#> [252,] 0.3329514 0.3289770 0.3380716
#> [253,] 0.3329357 0.3263634 0.3407010
#> [254,] 0.3321243 0.3262872 0.3415885
#> [255,] 0.3342013 0.3255493 0.3402494
#> [256,] 0.3336176 0.3257635 0.3406189
#> [257,] 0.3332930 0.3255765 0.3411304
#> [258,] 0.3347918 0.3215707 0.3436375
#> [259,] 0.3303894 0.3251828 0.3444278
#> [260,] 0.3336981 0.3277812 0.3385207
#> [261,] 0.3341566 0.3271959 0.3386476
#> [262,] 0.3328821 0.3275716 0.3395463
#> [263,] 0.3315802 0.3269365 0.3414833
#> [264,] 0.3330251 0.3268296 0.3401452
#> [265,] 0.3355148 0.3260143 0.3384709
#> [266,] 0.3389349 0.3360842 0.3249809
#> [267,] 0.3306336 0.3274291 0.3419372
#> [268,] 0.3343451 0.3269266 0.3387283
#> [269,] 0.3346456 0.3270158 0.3383385
#> [270,] 0.3329111 0.3271262 0.3399627
#> [271,] 0.3353357 0.3202605 0.3444039
#> [272,] 0.3342615 0.3250916 0.3406469
#> [273,] 0.3289777 0.3282204 0.3428019
#> [274,] 0.3328555 0.3278388 0.3393057
#> [275,] 0.3334183 0.3283081 0.3382736
#> [276,] 0.3361069 0.3307105 0.3331826
#> [277,] 0.3312520 0.3288663 0.3398816
#> [278,] 0.3338832 0.3255688 0.3405480
#> [279,] 0.3342974 0.3236412 0.3420614
#> [280,] 0.3347087 0.3195555 0.3457358
#> [281,] 0.3318220 0.3270395 0.3411385
#> [282,] 0.3369771 0.3294963 0.3335265
#> [283,] 0.3303411 0.3257557 0.3439032
#> [284,] 0.3366579 0.3307011 0.3326410
#> [285,] 0.3330937 0.3249491 0.3419571
#> [286,] 0.3314260 0.3253210 0.3432530
#> [287,] 0.3339969 0.3212776 0.3447254
#> [288,] 0.3337871 0.3253788 0.3408340
#> [289,] 0.3313790 0.3276281 0.3409929
#> [290,] 0.3302151 0.3265190 0.3432658
#> [291,] 0.3337998 0.3203435 0.3458567
#> [292,] 0.3358948 0.3230213 0.3410839
#> [293,] 0.3310928 0.3276170 0.3412903
#> [294,] 0.3348586 0.3233385 0.3418029
#> [295,] 0.3325888 0.3264189 0.3409923
#> [296,] 0.3335125 0.3273819 0.3391056
#> [297,] 0.3304676 0.3281143 0.3414181
#> [298,] 0.3318819 0.3270352 0.3410829
#> [299,] 0.3324472 0.3267860 0.3407668
#> [300,] 0.3337623 0.3276081 0.3386296